+++ lib/src/nes_ui.dart
@@ -1,2 +1,3 @@
+export 'controller.dart';
 export 'theme.dart';
 export 'widgets/widgets.dart';