+++ lib/src/widgets/widgets.dart
@@ -1,6 +1,7 @@
 export 'nes_button.dart';
 export 'nes_check_box.dart';
 export 'nes_container.dart';
+export 'nes_controller_focus.dart';
 export 'nes_icon.dart';
 export 'nes_input.dart';
 export 'nes_selection_list.dart';