+++ example/lib/gallery/sections/sections.dart
@@ -3,6 +3,7 @@ export 'checkboxes.dart';
 export 'containers.dart';
 export 'custom_extensions.dart';
 export 'icons.dart';
+export 'key_icons.dart';
 export 'selection_list.dart';
 export 'text_fields.dart';
 export 'texts.dart';