+++ lib/src/widgets/widgets.dart
@@ -4,4 +4,5 @@ export 'nes_container.dart';
 export 'nes_controller_focus.dart';
 export 'nes_icon.dart';
 export 'nes_input.dart';
+export 'nes_key_icon.dart';
 export 'nes_selection_list.dart';