+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # 0.2.0
 
-- feat: adding `NesSelectionList`
+- feat: add `NesSelectionList`
 - feat: `NesIcon` supports custom size
+- feat: add `NesIcons.gallery`
 
 # 0.1.0