+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 0.3.0
+
+- feat: add `backgroundColor` to `NesContainer`
+
 # 0.2.0
 
 - feat: add `NesSelectionList`
@@ -9,7 +13,10 @@
 - feat: add `NesIcons.redo`
 - feat: add `customExtensions`
 - feat: add keyboard support
-- feat: add `backgroundColor` to `NesContainer`
+- feat: add keyboard support 
+- feat: add `NesIcons.unpressedButton`
+- feat: add `NesIcons.pressedButton`
+- feat: add `NesKeyButton`
 
 # 0.1.0